pyproject.toml
[tool.poetry]
name = "nistai"
version = "0.1.0"
description = ""
authors = ["robomotic <paolo@robomotic.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.2"
uv = "^0.6.14"
djangorestframework = "^3.14.0"
pdfminer-six = "^20221105"
pillow = "^10.0.0"
python-docx = "^1.1.0"
pdfplumber = "^0.10.3"
pdf2image = "^1.17.0"
reportlab = "^4.1.0"
pymupdf = "^1.23.7"
uvicorn = "^0.34.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
